requirements.txt
streamlit
pypdf
sentence-transformers
transformers
faiss-cpu
torch